cardano: Moved BlockDate definitions to a separate module
